[#4] Minor code improvements
